# Remove allocations