Co-authored-by: Yingbo Ma <mayingbo5@gmail.com>

Before:
```julia
julia> @benchmark RecursiveFactorization.lu!(copyto!($B,$A),$p,Val(true),Val(false))
BenchmarkTools.Trial: 1488 samples with 1 evaluation.
 Range (min … max):  3.321 ms …  3.491 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.353 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.356 ms ± 21.492 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

      ▄▅▇█▁▂▂▃▅▆▄▁▂▃▄▅▂     ▅▅ ▂   ▁
  ▂▃▄▆██████████████████▇███████▇█▇█▆▅▇▄▃▇▅▄▃▄▄▃▃▃▂▃▂▃▃▁▂▂▃▂ ▅
  3.32 ms        Histogram: frequency by time        3.42 ms <

 Memory estimate: 112 bytes, allocs estimate: 4.

```
After:
```julia
julia> @benchmark RecursiveFactorization.lu!(copyto!($B,$A),$p,Val(true),Val(false))
BenchmarkTools.Trial: 1492 samples with 1 evaluation.
 Range (min … max):  3.311 ms …  3.482 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.345 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.346 ms ± 19.153 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

        ▃▁▅▂▃ ▄▄▃▄▅▃  ▁▄▄█▄▅▁▂▃▄▄▅▃▁
  ▂▁▂▄▅██████████████▇██████████████▄▅██▆▇▇▄▃▃▃▄▅▂▃▂▂▃▁▂▂▁▂▂ ▄
  3.31 ms        Histogram: frequency by time         3.4 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

The change is self evident.